Reorganized the code of data statistic helper functions
