+++ core/amber/src/main/scala/edu/uci/ics/texera/web/resource/dashboard/user/dataset/DatasetResource.scala
@@ -166,6 +166,32 @@ object DatasetResource {
     latestVersion
   }
 
+  // this function retrieves the total counts of dataset that belongs to the user
+  def getUserCreatedDatasetCount(uid: UInteger): Int = {
+    val count = context
+      .selectCount()
+      .from(DATASET)
+      .where(DATASET.OWNER_UID.eq(uid))
+      .fetchOne(0, classOf[Int])
+
+    count
+  }
+
+  // this function would return a list of dataset ids that belongs to the user
+  def getUserCreatedDatasetList(uid: UInteger): DatasetIDs = {
+    val datasetIDs = context
+      .select(
+        DATASET.DID
+      )
+      .from(DATASET)
+      .where(DATASET.OWNER_UID.eq(uid))
+      .fetch()
+
+    val idsList: List[UInteger] = datasetIDs.asScala.map(_.getValue(DATASET.DID)).toList
+
+    DatasetIDs(idsList)
+  }
+
   // this function create a new dataset version
   // the dataset is identified by did, the file changes/removals are contained in multiPart form
   // it returns the created dataset version if creation succeed, else return None
