+++ core/amber/src/main/scala/edu/uci/ics/texera/web/resource/dashboard/user/quota/UserQuotaResource.scala
@@ -2,24 +2,14 @@ package edu.uci.ics.texera.web.resource.dashboard.user.quota
 
 import edu.uci.ics.texera.web.SqlServer
 import edu.uci.ics.texera.web.auth.SessionUser
-import edu.uci.ics.texera.web.resource.dashboard.user.quota.UserQuotaResource.{
-  File,
-  MongoStorage,
-  Workflow,
-  deleteMongoCollection,
-  getUserAccessedFiles,
-  getUserAccessedWorkflow,
-  getUserCreatedDatasetCount,
-  getUserCreatedFile,
-  getUserCreatedWorkflow,
-  getUserMongoDBSize
-}
+import edu.uci.ics.texera.web.resource.dashboard.user.quota.UserQuotaResource.{File, MongoStorage, Workflow, deleteMongoCollection, getUserAccessedFiles, getUserAccessedWorkflow, getUserCreatedFile, getUserCreatedWorkflow, getUserMongoDBSize}
 import org.jooq.types.UInteger
 
 import java.util
 import javax.ws.rs._
 import javax.ws.rs.core.MediaType
 import edu.uci.ics.texera.web.model.jooq.generated.Tables._
+import edu.uci.ics.texera.web.resource.dashboard.user.dataset.DatasetResource.getUserCreatedDatasetCount
 import edu.uci.ics.texera.web.resource.dashboard.user.dataset.utils.DatasetStatisticsUtils.getUserDatasetSize
 import edu.uci.ics.texera.web.storage.MongoDatabaseManager
 import io.dropwizard.auth.Auth
@@ -38,12 +28,6 @@ object UserQuotaResource {
       description: String
   )
 
-  case class Dataset(
-      userId: UInteger,
-      datasetId: UInteger,
-      datasetName: String
-  )
-
   case class Workflow(
       userId: UInteger,
       workflowId: UInteger,
@@ -57,15 +41,7 @@ object UserQuotaResource {
       eid: UInteger
   )
 
-  def getUserCreatedDatasetCount(uid: UInteger): Int = {
-    val count = context
-      .selectCount()
-      .from(DATASET)
-      .where(DATASET.OWNER_UID.eq(uid))
-      .fetchOne(0, classOf[Int])
 
-    count
-  }
 
   def getCollectionName(result: String): String = {
